\name{lbiasgenerate}
\alias{lbiasgenerate}
\title{
Generates length-biased samples.}
\description{
This is a toy data simulator 
}
\usage{
lbiasgenerate(size)
}
\arguments{
  \item{size}{sample size.}
}

\details{
Add details here
}
\value{
\item{list1}{step function perhaps}
}
\references{
Add references here.
}
\author{
Pierre-Jerome and Vahid.
}
\note{
Add notes here about possible mistakes, misuse, etc.
}
\seealso{
\code{\link{nonpar.fit}}
}
\examples{
## Generates gamma length-biased data
survdata <- lbiasgenerate(size=1000)
}
\keyword{length-biased sampling}
