\name{par.fit}
\alias{par.fit}
\title{
Parametric fit}
\description{
Does Weibull, ... }
\usage{
par.fit(t, init, family = "weibull", extra = FALSE)
}
\arguments{
  \item{t}{
survival time matrix}
  \item{init}{
initial values for parameters}
  \item{family}{
parametric family to fit
}
  \item{extra}{
logical, if TRUE gives more information about parameter estimation}
}
\details{
structure is very similar to \code{par.fit} }
\value{
 \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
}
\references{
References about the parametric fit
}
\author{
Pierre-Jerome, Vahid.}
\seealso{
\link{par.fit}}
\examples{
# nothing 
}
\keyword{Weibull distribution}
\keyword{length-biased sampling}
